Just extract the sorting method as an utility.
