+++ src/components/Consent/computeConsentHash.js
@@ -10,6 +10,6 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { crc32, stableSerialization } from "../../utils/index.js";
+import { crc32, sortObjectKeysRecursively } from "../../utils/index.js";
 
-export default (obj) => crc32(stableSerialization(obj));
+export default (obj) => crc32(JSON.stringify(sortObjectKeysRecursively(obj)));
