+++ src/utils/index.js
@@ -60,4 +60,4 @@ export { default as uuid } from "./uuid.js";
 export { default as updateErrorMessage } from "./updateErrorMessage.js";
 export { default as validateIdentityMap } from "./validateIdentityMap.js";
 export { default as validateConfigOverride } from "./validateConfigOverride.js";
-export { default as stableSerialization } from "./stableSerialization.js";
+export { default as sortObjectKeysRecursively } from "./sortObjectKeysRecursively.js";
