--- src/utils/stableSerialization.js
@@ -1,36 +1,0 @@
-/*
-Copyright 2025 Adobe. All rights reserved.
-This file is licensed to you under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License. You may obtain a copy
-of the License at http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
-OF ANY KIND, either express or implied. See the License for the specific language
-governing permissions and limitations under the License.
-*/
-
-/**
- * Serializes a JavaScript object in a stable way.
- * It sorts the keys of objects. Arrays are not sorted.
- * @param {any} obj - The object to serialize
- * @returns {string} - The serialized string representation
- */
-const serialize = (obj) => {
-  if (Array.isArray(obj)) {
-    return obj.map((value) => serialize(value));
-  }
-
-  if (typeof obj === "object" && obj !== null) {
-    return Object.keys(obj)
-      .sort()
-      .reduce((memo, key) => {
-        memo[key] = serialize(obj[key]);
-        return memo;
-      }, {});
-  }
-
-  return obj;
-};
-
-export default (obj) => JSON.stringify(serialize(obj));
